fix(generation): text-length train/inference parity + stale-shift warning + CFG fp32
From a generation-pipeline review (follow-up to the 8cd55c0 audit):

- Text conditioning ran at a fixed 77-token zero-padded bucket at inference
  while training / training-eval / the reference pipeline all use natural
  token length (~12-20 for our prompts). use_attention_mask is False, so the
  ~60 pad tokens are fully attended — conditioning the model on input it
  never saw in LoRA training. Eager path now uses natural length (capped at
  cfg text_length like the reference); the 77/128 bucket is kept ONLY under
  torch.compile, where fixed shapes are required for graph reuse. The dead
  128 bucket could never trigger anyway (min(128, cap=77)).
- _flow_shift_override is sticky on the loaded model and silently survives
  removal of the ModelSampling node; both denoise loops now warn loudly when
  a non-default override is active.
- CFG combine upcasts to fp32 before chunk (reference-pipeline convention;
  the scheduler already integrates fp32) — applied to both production loops
  AND generate_eval_sample so eval/production stay in lockstep.
- a2a/inpaint/style-transfer stereo inputs now downmix by channel mean
  (training convention) instead of dropping the right channel.
- Fix vendored denoise_process trimming the channel axis instead of time
  (unused by ComfyUI nodes, but wrong).

+++ lora/train.py
@@ -712,7 +712,9 @@ def generate_eval_sample(model, dac_model, dataset_entry, device, dtype,
                 clip_feat=cfg_clip,
                 sync_feat=cfg_sync,
             )["x"]
-        v_uncond, v_cond = v_pred.chunk(2)
+        # CFG combine in fp32 (parity with production denoise loops; the
+        # scheduler integrates in fp32 internally)
+        v_uncond, v_cond = v_pred.float().chunk(2)
         v_guided = v_uncond + cfg_scale * (v_cond - v_uncond)
         latents = scheduler.step(v_guided, t, latents)[0]
